requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.44.0
django-crispy-forms==1.7.2
django-embed-video==1.3.3
django-tinymce==3.2.0
django-videofield==0.1.1
gunicorn==20.0.4
idna==2.10
jsmin==2.2.2
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.0
Pillow==8.0.0
pipe==1.6.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2020.1
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
whitenoise==5.2.0
